# Makefile: make 'make check' faster