--- a/cmd/explaintest/main.go
+++ b/cmd/explaintest/main.go
@@ -26,7 +26,7 @@ import (
 
 	"flag"
 	"github.com/go-sql-driver/mysql"
-	_ "github.com/mgechev/revive"
+	_ "github.com/mgechev/revive/fixtures"
 	"github.com/pingcap/errors"
 	"github.com/pingcap/parser/ast"
 	"github.com/pingcap/tidb/session"