@lysu 2018-12-19
@tiancaiamao please fix CI failure