@gregwebs 2018-12-19
I just tested this out. `make check` seems consistently a little bit slower for me than running directly. Running `revive` directly or via retool, but just not via the Makefile is equally fast by a little bit for me. Is the third change necessary? If so, we really need to understand why and find a better solution for managing our tooling.

```
make lint                                                                      18.78s user 0.69s system 516% cpu 3.766 total
revive -formatter friendly -config revive.toml                                 17.16s user 0.52s system 653% cpu 2.705 total
retool do revive -formatter friendly -config revive.toml                       17.80s user 0.46s system 659% cpu 2.769 total
```

Whenever possible in other projects I try to avoid creating a special environment inside a Makefile that only works in the Makefile. I don't know why the Makefile version is consistently slower.