@tiancaiamao 2018-12-20
@gregwebs 
I thought `make lint` on $(PACKAGES) is slow because it download packages from the network, but it seems I'm biased. 
And I'm not quite sure use  `./...` works.  Now I change `make lint` to work on $(FILES). 
Anyway, check $(PACKAGES) is quite slow, check $(FILES) is much better.

Before, on $(PACKAGES):
```
➜  tidb git:(make-check-accelerate) ✗ time make lint     
linting
make lint  25.51s user 28.43s system 74% cpu 1:12.65 total
```
After, on $(FILES)
```
➜  tidb git:(make-check-accelerate) ✗ time make lint
linting
make lint  5.00s user 1.02s system 166% cpu 3.609 total
```

After the change, `make lint` reports some errors for me:
```
linting
  ✘  https://revive.run/r#var-naming  var errTestSql should be errTestSQL  
  executor/set_test.go:53:2

  ✘  https://revive.run/r#var-naming  don't use underscores in Go names; var info_tables should be infoTables  
  infoschema/infoschema_test.go:250:2

  ✘  https://revive.run/r#var-naming  var stmtId should be stmtID  
  executor/prepared_test.go:87:3

  ✘  https://revive.run/r#superfluous-else  if block ends with a break statement, so drop this else and outdent its block  
  planner/core/prepare_test.go:262:10

  ✘  https://revive.run/r#var-naming  var globalAutoId should be globalAutoID  
  meta/autoid/autoid_test.go:63:2

  ✘  https://revive.run/r#var-naming  var globalAutoId should be globalAutoID  
  meta/autoid/autoid_test.go:154:2

  ✘  https://revive.run/r#var-naming  don't use underscores in Go names; var err_admin should be errAdmin  
  executor/executor_test.go:242:5

  ✘  https://revive.run/r#var-naming  don't use underscores in Go names; var create_time should be createTime  
  executor/show_test.go:624:2

  ✘  https://revive.run/r#var-naming  don't use underscores in Go names; var create_time should be createTime  
  executor/show_test.go:666:2

  ✘  https://revive.run/r#var-naming  var explainedSql should be explainedSQL  
  executor/merge_join_test.go:218:2

  ✘  https://revive.run/r#var-naming  var explainedSql should be explainedSQL  
  executor/merge_join_test.go:228:2

  ✘  https://revive.run/r#var-naming  don't use underscores in Go names; type mockPump_PullBinlogsClient should be mockPumpPullBinlogsClient  
  session/session_test.go:103:6

✘ 12 problems (12 errors, 0 warnings)
```

That sounds good.

-----------------

As for the retool, you know, the CI runs on a new clean environment every time, and it downloads files from the network.
It takes @iamxy a lot of effort to cache Go module in our own CI environment to skip the downloading.
If we put revive package under go.mod, we can benefit from @iamxy 's great work.