@gregwebs 2018-12-20
Tool usage workflows are unfortunately complex in go. We should continue the discussion about whether the retool approach is best. But this should be a separate PR or discussion.

revive should use the same system as the rest of the linting tools. So we should change how we use all of the tools or in a separate PR and keep revive under retool in this PR.


The current PR removes tool isolation in two ways 1) global install 2) potential conflict between tool dependencies with binary dependencies.
If we look at this problem a little more closely as a separate issue instead of rushing code through in this PR, I think we can come up with a better way.