@tiancaiamao 2018-12-21
I'm no discussing about the retool approach in this PR.
The purpose of this PR is clear: make 'make check' faster.

> The current PR removes tool isolation in two ways 
> 1) global install 
> 2) potential conflict between tool dependencies with binary dependencies.

No "global install" problem, because CI runs in a totally new clean environment, and the environment is destroyed after that. 
Maybe if I could put the revive binary in other place and it's more acceptable? 

I'll make a separate go.mod for tools binary, to address the dependencies pollution problem.
I promise to do it in the next PR, is it ok?
@gregwebs